ChatPrompt 語言載入+翻譯
